Refactored code according the requirements
