--- backend/contacts/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from contacts.models import ContactForm
-
-
-class ContactFormSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ContactForm
-        fields = "__all__"
