--- backend/contacts/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from .views import ContactFormView
-
-urlpatterns = [
-    path("", ContactFormView.as_view()),
-]
